* Fix build of m68k_disasm (after fixing a small warning)   - fix port to osx86   - fix issue about print_syscall * Add printing some more documentation   - Fix undo/redo documentation

+++ src/dbg/bp.c
@@ -34,7 +34,7 @@ int debug_bp(const char *str)
 	int hwbp = BP_NONE;
 	int num;
 
-	hwbp = config_get("dbg.hwbp");
+	hwbp = (int)config_get("dbg.hwbp");
 	switch(str[0]) {
 	case 's':
 		hwbp = BP_SOFT;
